+++ main.js
@@ -19,7 +19,8 @@ function createWindow () {
         width: 800,
         height: 600,
         resizable: false,
-        show: false
+        show: false,
+        title: "Arena Electron"
     });
 
     // and load the index.html of the app.
@@ -82,7 +83,9 @@ let createGame = (username, password) => {
         width: 650,
         height: 650,
         resizable: false,
-        show: false
+        show: false,
+        useContentSize: true,
+        title: "Arena Electron"
     });
 
     // socket = new WebSocket(`ws://${serverIp}:44444`);