+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.0.5
+- Implemented collisions
+  - Bullets bounce off blue and orange walls
+  - Players bounce off green and orange walls
+  - Red walls destroy bullets and hurt Players for 2HP/s
+  - Bullets damage Players other than the ones that shoot them
+  - Players in contact with other Players are hurt for 2HP/s
+  - Bullets are destroyed upon hitting a wall for the 4th time
+  - Bullet damage is inversely proportional to the number of times it bounces
+    - Max 10HP, Min 4HP
+
 # 0.0.4
 - Added Bullets in a basic form
   - Bullets do not get destroyed